Change onReset to use custombox
Included commented out code to point in the right direction. Once we make the HTML we can tie the modal's ID to the function and sprucen up it's display animation,and we can then tie a function to the "confirm" button in that modal that will fire the reset functions.